Underflow protection

+++ Game.sol
@@ -192,7 +192,7 @@ contract CastlediceGame {
                     if (column + 1 < FIELD_WIDTH && isGood[row][column+1]) {
                         needToRemove = false;
                     }
-                    if (column - 1 >= 0 && row + 1 < FIELD_HEIGHT && isGood[row+1][column-1]) {
+                    if (int256(column) - 1 >= 0 && row + 1 < FIELD_HEIGHT && isGood[row+1][column-1]) {
                         needToRemove = false;
                     }
 
@@ -217,16 +217,16 @@ contract CastlediceGame {
             for (uint256 column = bluePlayerStart.column+1; column < FIELD_WIDTH; column++) {
                 if (room.boardState[row][column] == BoardState.BLUE) {
                     bool needToRemove = true;
-                    if (row - 1 >= 0 && isGood[row-1][column]) {
+                    if (int256(row) - 1 >= 0 && isGood[row-1][column]) {
                         needToRemove = false;
                     }
-                    if (column - 1 >= 0 && row - 1 >= 0 && isGood[row-1][column-1]) {
+                    if (int256(column) - 1 >= 0 && row - 1 >= 0 && isGood[row-1][column-1]) {
                         needToRemove = false;
                     }
-                    if (column - 1 >= 0 && isGood[row][column-1]) {
+                    if (int256(column) - 1 >= 0 && isGood[row][column-1]) {
                         needToRemove = false;
                     }
-                    if (column + 1 < FIELD_WIDTH && row - 1 >= 0 && isGood[row-1][column+1]) {
+                    if (column + 1 < FIELD_WIDTH && int256(row) - 1 >= 0 && isGood[row-1][column+1]) {
                         needToRemove = false;
                     }
